updated the example of modularize-imports to utilize the App Router
