+++ examples/modularize-imports/app/_components/halves/LeftHalf.tsx
