+++ examples/modularize-imports/app/_components/halves/index.tsx
