--- examples/modularize-imports/pages/index.tsx
@@ -1,13 +1,0 @@
-// @ts-ignore
-import { LeftHalf, RightHalf } from "../components/halves";
-// @ts-ignore
-import { HeadingLarge } from "../components/ui";
-
-export default function Index() {
-  return (
-    <HeadingLarge>
-      <LeftHalf />
-      <RightHalf />
-    </HeadingLarge>
-  );
-}
